packages/ingestion/pyproject.toml
[project]
name = "pnseg-ingestion"
version = "0.1.0"
dependencies = [
    "pnseg-core",
    "kaggle",
    "python-dotenv",
    "joblib",
    "fastapi",
    "uvicorn"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["ingestion*"]